If hostname part of an address is empty, assume localhost.
Binding plugins to a TCP port on localhost using ServeTCP("foo", ":12345")
does not work on Windows. This is because the golang net package does
not handle empty hostname/IP. https://github.com/golang/go/issues/6290
tracks it and is slated to be fixed in Go 1.8.

This change explicitly adds "localhost" if there's no hostname and
serve as a workaround until Golang fixes it.

Signed-off-by: Anusha Ragunathan <anusha@docker.com>