Remove use of cgo
This makes it possible to cross-compile plugins.
Builds on top of docker/go-connections#26

If a plugin author wants to do group lookup, they can do it in their own
way and pass in the gid.
As is, this is too much for a library and makes it incredibly annoying
to use.

Signed-off-by: Brian Goff <cpuguy83@gmail.com>