Refactor plugin handlers to allow composable plugins.
This patch refactors the plugin handlers so they implement http.Handler, and
exposes the individual initMux methods as public methods. This allows an
advanced user to combine multiple drivers and plugin handlers under a common
mux - enabling the likely use-case of a network driver and IPAM driver which
should be launched together.

This patch is mostly backwards compatible with the existing API with the
exception of sdk.Handler.NewHandler() which by necessity is adapted to provide
for step-wise addition of implementations.